PR to fix numpy TypeErrors from indexing arrays with non-integers

+++ MOSFIRE/Flats.py
@@ -375,7 +375,7 @@ def find_edge_pair(data, y, roi_width, edgeThreshold=450):
     '''
 
     def select_roi(data, roi_width):
-        v = data[y-roi_width:y+roi_width, xp-2:xp+2]
+        v = data[int(y-roi_width):int(y+roi_width), int(xp)-2:int(xp)+2]
         v = np.median(v, axis=1) # Axis = 1 is spatial direction
 
         return v
@@ -421,8 +421,8 @@ def find_edge_pair(data, y, roi_width, edgeThreshold=450):
 
             between = offset + width/2
             if 0 < between < len(v)-1:
-                start = np.max([0, between-2])
-                stop = np.min([len(v),between+2])
+                start = int(np.max([0, between-2]))
+                stop = int(np.min([len(v),between+2]))
                 yposs_bot_scatters.append(np.min(v[start:stop])) # 5
 
                 if False:
@@ -791,7 +791,7 @@ def find_and_fit_edges(data, header, bs, options,edgeThreshold=450):
     xposs_top_this = np.arange(10,2000,100)
     yposs_top_this = topfun(xposs_top_this)
 
-    initial_edges = np.array([2034])
+    initial_edges = np.array([2034], dtype=np.int)
     edge = 2034
 
     # build an array of values containing the lower edge of the slits
@@ -799,14 +799,14 @@ def find_and_fit_edges(data, header, bs, options,edgeThreshold=450):
     for target in xrange(len(ssl)):        
     # target is the slit number
         edge -= DY * numslits[target]
-        initial_edges=np.append(initial_edges,edge)
+        initial_edges=np.append(initial_edges,int(edge))
 
     # collapse the 2d flat along the walenegth axis to build a spatial profile of the slits
     vertical_profile = np.mean(data, axis=1)
 
     # build an array containing the spatial positions of the slit centers, basically the mid point between the expected
     # top and bottom values of the slit pixels
-    spatial_centers = np.array([])
+    spatial_centers = np.array([], dtype=np.int)
     for k in np.arange(0,len(initial_edges)-1):
         spatial_centers = np.append(spatial_centers,(initial_edges[k]+initial_edges[k+1])/2)
     #slit_values=np.array([])
